+++ apps/expo-go/ios/Exponent/Kernel/Core/EXKernel.m
@@ -2,13 +2,13 @@
 
 #import "EXAppState.h"
 #import "EXAppViewController.h"
-#import "EXBuildConstants.h"
 #import "EXKernel.h"
+
+#import "Expo_Go-Swift.h"
 #import "EXAbstractLoader.h"
 #import "EXKernelAppRecord.h"
 #import "EXKernelLinkingManager.h"
 #import "EXLinkingManager.h"
-#import "EXVersions.h"
 #import "EXKernelDevKeyCommands.h"
 
 #import <EXConstants/EXConstantsService.h>
@@ -131,13 +131,6 @@ NSString * const kEXReloadActiveAppRequest = @"EXReloadActiveAppRequest";
 - (void)_postNotificationName: (NSNotificationName)name
 {
   [[NSNotificationCenter defaultCenter] postNotificationName:name object:nil];
-}
-
-#pragma mark - App props
-
-- (nullable NSDictionary *)initialAppPropsFromLaunchOptions:(NSDictionary *)launchOptions
-{
-  return nil;
 }
 
 #pragma mark - App State