+++ apps/native-component-list/src/screens/UI/DateTimePickerScreen.android.tsx
@@ -1,5 +1,10 @@
-import { DateTimePicker, DateTimePickerProps, Picker, Column } from '@expo/ui/jetpack-compose';
-import { Host } from '@expo/ui/swift-ui';
+import {
+  DateTimePicker,
+  DateTimePickerProps,
+  Picker,
+  Column,
+  Host,
+} from '@expo/ui/jetpack-compose';
 import * as React from 'react';
 import { ScrollView, Text } from 'react-native';