+++ packages/expo-router/src/split-view/split-view.tsx
@@ -1,9 +1,5 @@
 import React, { createContext, isValidElement, use, type ReactNode } from 'react';
-import {
-  SplitViewHost,
-  SplitViewScreen,
-  type SplitViewHostProps,
-} from 'react-native-screens/experimental';
+import { Split, type SplitHostProps } from 'react-native-screens/experimental';
 
 import { SplitViewColumn, SplitViewInspector } from './elements';
 import { IsWithinLayoutContext } from '../layouts/IsWithinLayoutContext';
@@ -12,9 +8,9 @@ import { Slot } from '../views/Navigator';
 const IsWithinSplitViewContext = createContext(false);
 
 /**
- * For full list of supported props, see [`SplitViewHostProps`](https://github.com/software-mansion/react-native-screens/blob/main/src/components/gamma/split-view/SplitViewHost.types.ts#L124)
+ * For full list of supported props, see [`SplitHostProps`](http://github.com/software-mansion/react-native-screens/blob/main/src/components/gamma/split/SplitHost.types.ts#L117)
  */
-export interface SplitViewProps extends Omit<SplitViewHostProps, 'children'> {
+export interface SplitViewProps extends Omit<SplitHostProps, 'children'> {
   children?: ReactNode;
 }
 
@@ -68,13 +64,13 @@ function SplitViewNavigator({ children, ...splitViewHostProps }: SplitViewProps)
 
   // The key is needed, because number of columns cannot be changed dynamically
   return (
-    <SplitViewHost key={numberOfSidebars + numberOfInspectors} {...splitViewHostProps}>
+    <Split.Host key={numberOfSidebars + numberOfInspectors} {...splitViewHostProps}>
       {columnChildren}
-      <SplitViewScreen.Column>
+      <Split.Column>
         <WrappedSlot />
-      </SplitViewScreen.Column>
+      </Split.Column>
       {inspectorChildren}
-    </SplitViewHost>
+    </Split.Host>
   );
 }