+++ packages/expo-server/src/rendering.ts
@@ -1,4 +1,5 @@
-import { GetStaticContentOptions } from './manifest';
+import { type ImmutableRequest } from './ImmutableRequest';
+import { type GetStaticContentOptions } from './manifest';
 
 /**
  * The SSR render module exported from `_expo/server/render.js`.
@@ -16,7 +17,10 @@ export interface RenderOptions {
 
 /** Module exported from loader bundle, typically `_expo/loaders/[ROUTE].js` */
 export interface LoaderModule {
-  loader?(args: { params: Record<string, string>; request: Request }): Promise<unknown> | unknown;
+  loader(args: {
+    params: Record<string, string>;
+    request: ImmutableRequest;
+  }): Promise<unknown> | unknown;
 }
 
 export type SsrRenderFn = (request: Request, options?: RenderOptions) => Promise<string>;